ml/requirements.txt
# ============================================
# CrisisTriage AI - ML Dependencies
# ============================================

# --- Deep Learning ---
# Install PyTorch based on your system:
# CPU: pip install torch torchvision torchaudio
# CUDA: pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu121
torch>=2.1.0
torchvision>=0.16.0
torchaudio>=2.1.0

# --- NLP / Transformers ---
transformers>=4.35.0
tokenizers>=0.15.0
sentence-transformers>=2.2.0  # For semantic embeddings

# --- Audio Processing ---
librosa>=0.10.0
soundfile>=0.12.0
audioread>=3.0.0
# opensmile>=2.5.0  # Optional: Extended prosody features

# --- Speech Recognition ---
openai-whisper>=20231117
# faster-whisper>=0.9.0  # Alternative: Optimized Whisper

# --- Data Processing ---
numpy>=1.24.0
pandas>=2.1.0
scipy>=1.11.0
scikit-learn>=1.3.0

# --- Experiment Tracking ---
mlflow>=2.9.0
# wandb>=0.16.0  # Alternative: Weights & Biases

# --- Configuration ---
hydra-core>=1.3.0
omegaconf>=2.3.0

# --- Explainability ---
shap>=0.44.0
captum>=0.6.0  # PyTorch interpretability

# --- Visualization ---
matplotlib>=3.8.0
seaborn>=0.13.0
plotly>=5.18.0

# --- Jupyter ---
jupyter>=1.0.0
ipywidgets>=8.1.0

# --- Utilities ---
tqdm>=4.66.0
python-dotenv>=1.0.0
pyyaml>=6.0.0

# --- Testing ---
pytest>=7.4.0
pytest-cov>=4.1.0

# --- Code Quality ---
black>=23.11.0
isort>=5.12.0
flake8>=6.1.0
mypy>=1.7.0
